+++ app/controllers/Home.php
@@ -12,9 +12,10 @@ class Home extends Controller
   public function tambahObat()
   {
     $data['title'] = 'tambah obat';
-    $data['obat'] = $this->getData();
-    $data['errors'] = $this->getError();
-    $data['format']=$this->getFormatDefault();
+    $validator = $this->validate('Validate');
+    $data['obat'] = $validator->getData();
+    $data['errors'] = $validator->getError();
+    $data['format']=$validator->getFormatDefault();
     $this->view('templates/header', $data);
     $this->view('home/tambahObat', $data);
     $this->view('templates/footer');
@@ -31,9 +32,10 @@ class Home extends Controller
   {
     $data['obat'] = $this->model('Obat_model')->getObatByKode($kode);
     $data['title'] = 'update ' . $data['obat']['nama_generik'] . ' / ' . $data['obat']['nama_merek'];
-    $this->setData($data);
-    $data['errors'] = $this->getError();
-    $data['format']=$this->getFormatDefault();
+    $validator = $this->validate('Validate');
+    $validator->setData($data);
+    $data['errors'] = $validator->getError();
+    $data['format']=$validator->getFormatDefault();
     $this->view('templates/header', $data);
     $this->view('home/update', $data);
     $this->view('templates/footer');
@@ -42,8 +44,9 @@ class Home extends Controller
   {
     $data['title'] = 'tambah obat';
     $data['obat'] = $_POST;
-    $this->setData($data['obat']);
-    if($this->validasi()){
+    $validator = $this->validate('Validate');
+    $validator->setData($data['obat']);
+    if($validator->validasi()){
       if ($this->model('Obat_model')->tambahDataObat($data['obat']) > 0) {
         Flasher::setFlash('berhasil', 'ditambahkan', 'success', 'obat');
         header('Location: ' . BASEURL . '/home');
@@ -54,9 +57,9 @@ class Home extends Controller
         exit;
       }
     } else {
-      $data['errors'] = $this->getError();
+      $data['errors'] = $validator->getError();
     }
-    $data['format'] = $this->getFormat();
+    $data['format'] = $validator->getFormat();
     $this->view('templates/header', $data);
     $this->view('home/tambahObat', $data);
     $this->view('templates/footer');
@@ -78,8 +81,9 @@ class Home extends Controller
     $data['obat'] = $_POST;
     $tempDatabase = $this->model('Obat_model')->getObatByKode($data['obat']['kode']);
     $data['title'] = 'update ' . $tempDatabase['nama_generik'] . ' / ' . $tempDatabase['nama_merek'];
-    $this->setData($data['obat']);
-    if($this->validasi()){
+    $validator = $this->validate('Validate');
+    $validator->setData($data['obat']);
+    if($validator->validasi()){
       if ($this->model('Obat_model')->ubahDataObat($data['obat']) > 0) {
         Flasher::setFlash('berhasil', 'diubah', 'success', 'obat');
         header('Location: ' . BASEURL . '/home');
@@ -90,9 +94,9 @@ class Home extends Controller
         exit;
       }
     } else {
-      $data['errors'] = $this->getError();
+      $data['errors'] = $validator->getError();
     }
-    $data['format'] = $this->getFormat();
+    $data['format'] = $validator->getFormat();
     $this->view('templates/header', $data);
     $this->view('home/update', $data);
     $this->view('templates/footer');
